Changed read to accommodate different PCLs
The read vi used to discard all data if there was a PCL mismatch such that the reader was slower than the writer. It now reads all multiples of the data from the buffer and just uses the last complete dataset
The init routine has been modified to increase the buffer size. A ratio of 10:1 between the PCLs can be accommodated with the current settings.